# doc/api.yaml
openapi: "3.1.0"
info:
  title: Arsmedia Backend API
  description: >
    Backend API for the **Arsmedia Application**. This API provides endpoints for:
    
    **Video Capabilities**
    - Convert between formats: .avi → .mp4, .mkv → .mov, .webm → .avi, etc.
    - Change codecs (H.264, H.265, VP9, AV1).
    - Extract or replace audio tracks.
    - Cut, trim, merge, or concatenate videos without re-encoding.
    - Resize, crop, rotate, or flip videos.
    - Add subtitles or burn them into the video.
    - Apply filters: blur, sharpen, grayscale, watermark.
    - **Video compression**

    **Audio Capabilities**
    - Convert between audio formats: .mp3 → .wav, .flac → .aac, etc.
    - Change bitrate, sample rate, or channels.
    - Extract audio from a video file.
    - Normalize or adjust volume.
    - Apply effects: fade in/out, echo, equalizer.
    - Mix multiple audio tracks.

    **Image Capabilities**
    - Convert between image formats: .png → .jpg, .bmp → .tiff, etc.
    - Extract frames from a video.
    - Create a video or GIF from images.
    - Resize, crop, or apply filters.
    - Generate thumbnails.

    **Authentication & Security**
    - **Client Authentication**: All non-public endpoints require X-CLIENT-ID and X-CLIENT-SECRET headers for client verification
    - **User Authentication**: Protected endpoints require JWT Bearer token obtained from /auth/token/token-pairs
...
    - Local requests (localhost/127.0.0.1) bypass client secret validation

...

...
  contact:
...
servers:
...
paths:
...
  /api/media/videos/compression-status/{jobId}:
    get:
      operationId: getCompressionStatus
      tags:
        - Video
      summary: Get compression job status by job ID
      description: >
        Retrieves the current status and details of a specific compression job by its unique job ID.
        **Requires client credentials and Bearer token authentication.**
      security:
        - ClientSecret: [ ]
          BearerAuth: [ ]
      parameters:
        - name: jobId
          in: path
          required: true
          description: Unique identifier of the compression job
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Compression job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompressionJobStatusResponse"
        "400":
          description: Invalid request - jobId is blank or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "jobId parameter is required and cannot be blank"
                path: "/api/media/videos/compression-status/"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "404":
          description: Compression job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 404
                error: "Not Found"
                message: "Compression job with ID 'job-123' not found"
                path: "/api/media/videos/compression-status/job-123"
                errorCode: "ENTITY_NOT_FOUND"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/media/videos/compression-jobs/{videoId}:
    get:
      operationId: getCompressionJobsByVideoId
      tags:
        - Video
      summary: Get all compression jobs for a video
      description: >
        Retrieves a list of all compression jobs associated with a specific video ID.
        **Requires client credentials and Bearer token authentication.**
      security:
        - ClientSecret: [ ]
          BearerAuth: [ ]
      parameters:
        - name: videoId
          in: path
          required: true
          description: Unique identifier of the video
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: List of compression jobs retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompressionJobStatusResponse"
        "400":
          description: Invalid request - videoId is blank or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "videoId parameter is required and cannot be blank"
                path: "/api/media/videos/compression-jobs/"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "404":
          description: Video not found or no compression jobs found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 404
                error: "Not Found"
                message: "No compression jobs found for video with ID 'video-123'"
                path: "/api/media/videos/compression-jobs/video-123"
                errorCode: "ENTITY_NOT_FOUND"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/media/videos/compress/{userEmail}:
    post:
      operationId: compressVideo
      tags:
        - Video
      summary: Compress a video
      description: >
        Initiates video compression process and returns job status with tracking information.
        **Requires client credentials and Bearer token authentication.**
      security:
        - ClientSecret: [ ]
          BearerAuth: [ ]
      parameters:
        - name: userEmail
          in: path
          required: true
          description: Email of the user who owns the video
          schema:
            type: string
            format: email
        - name: bucket_key
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: The storage bucket key of the video to compress
      responses:
        "200":
          description: Video compression initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompressionJobStatusResponse"
        "400":
          description: Invalid request - missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "bucket_key parameter is required and cannot be empty"
                path: "/api/media/videos/compress/user@example.com"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 404
                error: "Not Found"
                message: "Video with bucket key 'video-123.mp4' not found"
                path: "/api/media/videos/compress/user@example.com"
                errorCode: "ENTITY_NOT_FOUND"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/media/videos/compressed:
    get:
      operationId: getAllCompressedVideosOfUser
      tags:
        - Video
      summary: Get all compressed videos for a user
      description: >
        Retrieves a list of all compressed videos belonging to a specific user.
        **Requires client credentials and Bearer token authentication.**
      security:
        - ClientSecret: [ ]
          BearerAuth: [ ]
      parameters:
        - name: userEmail
          in: query
          required: true
          description: Email of the user to retrieve compressed videos for
          schema:
            type: string
            format: email
      responses:
        "200":
          description: List of compressed videos retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "400":
          description: Invalid request - missing or invalid userEmail parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "userEmail parameter is required and must be a valid email"
                path: "/api/media/videos/compressed"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    ClientSecret:
      type: apiKey
      in: header
      name: X-CLIENT-SECRET
      description: >
        Client secret for API access. Must be provided along with X-CLIENT-ID header.
        Local requests (localhost/127.0.0.1) bypass this requirement.
        Contact API administrator to obtain client credentials.

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from /auth/token/token-pairs after login.
        Include in Authorization header as: `Bearer <token>`

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key obtained from /auth/api-keys/generate.
        Required for admin endpoints like /users.
        Include in X-API-Key header.
  responses:
    Unauthorized:
      description: Authentication required - missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 401
            error: "Unauthorized"
            message: "Authentication required"
            path: "/api/endpoint"
            errorCode: "MISSING_AUTHORIZATION"
...
    InvalidClientCredentials:
      description: Invalid or missing client credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 403
            error: "Forbidden"
            message: "Invalid or missing client credentials"
            path: "/api/endpoint"
            errorCode: "AUTHORIZATION_DENIED"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            path: "/api/endpoint"
            errorCode: "INTERNAL_SERVER_ERROR"
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP status reason phrase
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message
          example: "Validation failed"
        path:
          type: string
          description: Request path where error occurred
          example: "/path/of/the/endpoint"
        errorCode:
          type: string
          description: Application-specific error code
          enum:
            - AUTHORIZATION_DENIED
            - ENTITY_NOT_FOUND
            - CONSTRAINT_VIOLATION_ON_FIELDS
            - TOKEN_UNIQUENESS_CONFLICT
            - BAD_FORM
            - MISSING_AUTHORIZATION
            - INVALID_AUTHORIZATION_FORMAT
            - INVALID_TOKEN
            - TOKEN_NOT_FOUND
            - USER_DEACTIVATED
            - AUTHENTICATION_FAILED
            - INTERNAL_SERVER_ERROR
            - MISSING_REQUIRED_PARAMETER
            - INVALID_PARAMETER
            - MALFORMED_JSON
            - METHOD_NOT_ALLOWED
      required:
        - timestamp
        - status
        - error
        - message
        - path
    CompressionJobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Unique identifier for the compression job
        status:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - FAILED
            - CANCELLED
          description: Current status of the compression job
        createdAt:
          type: string
          format: date-time
          description: When the compression job was created
        completedAt:
          type: string
          format: date-time
          description: When the compression job was completed (if finished)
        compressedVideoId:
          type: string
          description: ID of the compressed video (when compression is successful)
        compressedVideoUrl:
          type: string
          description: URL to access the compressed video (when compression is successful)
        errorMessage:
          type: string
          description: Error message if compression failed
        attemptCount:
          type: integer
          format: int32
          description: Number of compression attempts made
      required:
        - jobId
        - status
        - createdAt
        - attemptCount
...
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        pseudo:
          type: string
        phoneNumber:
          type: string
        imageProfileBucketKey:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        password:
          type: string
          format: password
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        username:
          type: string
        authorities:
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
        activated:
          type: boolean
        enabled:
          type: boolean
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
        credentialsNonExpired:
          type: boolean
...
    Video:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        size:
          type: number
          format: double
        sizeType:
          type: string
          enum:
            - BYTES
            - KB
            - MB
            - GB
            - TB
            - PB
            - EB
        fileType:
          type: string
          enum:
            - VIDEO
            - IMAGE
            - AUDIO
        createdAt:
          type: string
          format: date-time
        filePath:
          type: string
          description: Path to the file in storage
        owner:
          $ref: "#/components/schemas/User"
        duration:
          type: number
          format: double
          description: Video duration in seconds
        codec:
          type: string
          enum:
            - H264
            - H265
            - MPEG1
            - MPEG2
            - MPEG4
            - VP8
            - VP9
            - AV1
            - THEORA
            - WMV1
            - WMV2
            - WMV3
            - MJPEG
            - PRORES
            - DNXHD
            - CINEFORM
            - DV
            - XVID
            - DIVX
            - HEVC
            - UNKNOWN
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        frameRate:
          type: number
          format: double
        aspectRatio:
          type: string
        containerFormat:
          type: string
          enum:
            - MP4
            - MKV
            - MOV
            - AVI
            - FLV
            - WMV
            - WEBM
            - MPEG_TS
            - MPEG_PS
            - THREEGP
            - OGG
            - M4A
            - WAV
            - FLAC
            - ASF
            - APE
            - AIFF
            - UNKNOWN
        bitRate:
          type: number
          format: double
        audioCodec:
          type: string
          enum:
            - NONE
            - AAC
            - MP3
            - MP2
            - AC3
            - EAC3
            - DTS
            - PCM
            - PCM_S16LE
            - PCM_S24LE
            - FLAC
            - OPUS
            - VORBIS
            - WMA
            - ALAC
            - AMR_NB
            - AMR_WB
            - G711
            - G722
            - UNKNOWN
        audioChannels:
          type: integer
          format: int32
        audioSampleRate:
          type: integer
          format: int32
...
security:
  - BearerAuth: [ ]